djnews/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
django-filter==21.1
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.10.0
Markdown==3.3.6
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
zipp==3.7.0
